Reformat info files and remove pandoc dependency
The reliance on Markdown was just a stylistic choice to match the
majority of projects on Github ... and my own preferences. Since the
reliance on pandoc to transform the README for releases was causing
issues (and apparently not working anyway) it makes sense to just get
rid of it.

Fixes #4
Fixes #5