Fail DNS tests when the query times out or fails
In the rare event that the DNS servers time out or no nameserver returns
an answer to a DNS query, we handle the error state returned for the DNS
check.

I made the decision to just add two more failure diagnoses for this
purpose. The alterative would be to try the query again, to see if an
answer will be returned on the retry. However, retrying isn't favorable
because:

1. When no nameservers return a response to a query, this indicates
a misconfiguration on the server running the check. That type of failure
isn't transient; it requires reconfiguration, so we shouldn't retry the
request.

2. `dnspython` times out a DNS request after 30 seconds. Retrying the
query could cause _another_ 30 second timeout to occur. Waiting a full
minute for a response is a terrible user experience, so we should make
the developer handle this case.

Fixes #3