Switch to use Hatch to manage the project
Hatch seems like it has enough mind share and compatibility that it
makes sense to switch to it as a project management tool. It reduces the
number of configuration files to one, can manage multiple Pythons and
multiple execution environments, and generally has all of the
affordances I feel that we need to manage things.

This change was spurred by my needing to look up once again how to
release the library. This was the last straw because I don't spend much
time working in Python these days and can't remember how to do it every
time I want to release a new version.

This should be a major quality of life improvement for anyone wanting to
work on the project.