Fix smoke test to show the packaging error
The original test didn't show it because Python was smartly loading the
files from the current directory. This change makes the virtualenv in
a temporary directory, installs the wheel, and runs the smoke test from
that context. This means it cannot pick up on the local files anymore.

This should lead to a failed test. If so, I will adjust the project file
to fix the issue.