Move to src layout and fix build process
The 2.0.0 release was botched due to a packaging error where the package
directory was misnamed as `pyIsEmail` instead of the expected
`pyisemail`. This was due to Python automatically adding the current
directory to the `PYTHONPATH` when invoked and my time of being away
from Python and forgetting that.

This change reorganizes the repository around the [src layout][1] which
prevents the issue since an `import pyisemail` will no longer load from
the current path. It also changes the test setup to install the package
in the environment instead of using an editable install. This way we're
actually using the built wheel as the source of the packages, which
verifies that the wheel works as expected.

[1]: https://blog.ionelmc.ro/2014/05/25/python-packaging/